Se actualiza el servicio de Album, además se actualizan los modelos y las respuestas para que el ID de cada entidad sea un Long en lugar de un Integer.

+++ src/main/java/ex_grupo11/ex_api_grupo11/models/Album.java
@@ -22,7 +22,7 @@ import lombok.NoArgsConstructor;
 public class Album {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idAlbum;
+    private Long idAlbum;
 
     @Column(nullable = false, length = 100)
     private String nombreAlbum;
