+++ src/main/java/ex_grupo11/ex_api_grupo11/responses/UsuarioAlbumResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class UsuarioAlbumResponse {
-    private Integer idUsuarioAlbum;
+    private Long idUsuarioAlbum;
     private UsuarioResponse usuario;
     private AlbumResponse album;
 }
